backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
python-docx
PyPDF2
openpyxl
